fix: update namespace query
WIP: [
