+++ queries/go/namespace.scm
@@ -3,7 +3,8 @@
   arguments: (argument_list . (interpreted_string_literal) @namespace.name))
   (#any-of? @func_name
     "Describe" "FDescribe" "PDescribe" "XDescribe"
-    "DescribeTable" "FDescribeTable" "PDescribeTable" "XDescribeTable"
     "Context" "FContext" "PContext" "XContext"
     "When" "FWhen" "PWhen" "XWhen"
+    "DescribeTable" "FDescribeTable" "PDescribeTable" "XDescribeTable"
+    "DescribeTableSubtree" "FDescribeTableSubtree" "PDescribeTableSubtree" "XDescribeTableSubtree"
   )) @namespace.definition
